[fix] s3에 저장된 url 리턴받아 GoodsImage 테이블에 저장하도록 수정
